consertei a strncmp e mudei a tipagem na lib

+++ ft_strncmp.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-size_t	ft_strncmp(char *s1, char *s2, size_t n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
 	size_t i;
